test: cover planning-profile drafts, industry-template handlers, calendar-selection official-override
Companion tests for the Klacks.Api commit of the same name — planning-profile
skill/handler tests, settings value validator, calendar-selection
official-override mapper/service tests, scheduling-rule repository/read-skill
updates, and active-industries/region-setup provider test updates.

Full suite green: 6762/6762 passed.

+++ Skills/SchedulingRuleReadSkillsTests.cs
@@ -82,7 +82,7 @@ public class SchedulingRuleReadSkillsTests
         data.GetProperty("MaxDailyHours").GetDouble().ShouldBe(SchedulingPolicyDefaults.MaxDailyHours);
         data.GetProperty("MaxConsecutiveDays").GetInt32().ShouldBe(SchedulingPolicyDefaults.MaxConsecutiveDays);
         data.GetProperty("MaxWeeklyHours").GetDouble().ShouldBe(SchedulingPolicyDefaults.MaxWeeklyHours);
-        data.GetProperty("MinRestDays").GetInt32().ShouldBe(SchedulingPolicyDefaults.MinRestDays);
+        data.GetProperty("MinRestDays").GetDecimal().ShouldBe(SchedulingPolicyDefaults.MinRestDays);
         data.GetProperty("IsFallback").GetBoolean().ShouldBeTrue();
     }
 }
